+++ src/rooms/room_details.rs
@@ -1,0 +1,3 @@
+//! Different versions of the room details endpoint.
+
+pub mod v1;