+++ src/users/list_joined_rooms/v1.rs
@@ -10,7 +10,7 @@ const METADATA: Metadata = metadata! {
     rate_limited: false,
     authentication: AccessToken,
     history: {
-        unstable => "/_synapse/admin/v1/users/:user_id/joined_rooms",
+        unstable => "/_synapse/admin/v1/users/{user_id}/joined_rooms",
     }
 };