+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Improvement:
 
 * The list_room response now includes the `room_type` field
 * Add room_details api
+* Add room_members api
 
 # 0.7.0