+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Breaking changes:
 
 * The list_room response changes the fields `version`, `join_rules`, `guest_access` and `history_visibility` to be an option
 * The list_room response changes the `join_rules` field to be `Option<SpaceRoomJoinRule>`
+* `background_update::run::v1::JobName` is now non-exhaustive.
+* `RoomSortOrder` and `RoomDirection` in `rooms::list_rooms::v1` are now non-
+  exhaustive. Their `PartialOrd` and `Ord` implementations now use their string
+  representation instead of the order in which they are defined in the enum.
 
 Improvement: